Fixed error in `pull` method of `Todo` model class
- Fix-1: Previously `pull` method was not taking any argument, but
we need to pass `self` as an argument as pull is an `instance method`.

- Fix-2: Previously `pull` was returning a whole new `Todo` model
object when all the fields of Todo is present, but this was against
the working philosophy `pull` method. Pull should retrieve data from
`firestore`(only, if necessary) and assign those data to its own
fiedls. That's what the `pull` method should do. So, the fix was just
to return self rather than retunring a whole new object.